doing view here
