+++ MacFlashCard/ContentView.swift
@@ -1,0 +1,25 @@
+//
+//  ContentView.swift
+//  MacFlashCard
+//
+//  Created by Olivier Lambert Rouillard on 2023-09-05.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        ZStack {
+            Text("Something")
+        }
+//        iPhoneView()
+//            .frame(minWidth: 700, minHeight: 300)
+    }
+}
+
+struct MacContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+//            .environmentObject(FlashCard())
+    }
+}
